+++ src/opcode.rs
@@ -103,22 +103,22 @@ pub fn shamt(instr: u32) -> u32 {
     return (imm_I(instr) & 0x1f) as u32;
 }
 
-pub fn csr(instr: u32) -> u64 {
+pub fn csr(instr: u32) -> u32 {
     // csr[11:0] = inst[31:20]
-    return ((instr & 0xfff00000) >> 20) as u64;
+    return ((instr & 0xfff00000) >> 20) as u32;
 }
 
-pub fn imm_B(instr: u32) -> u64 {
+pub fn imm_B(instr: u32) -> u32 {
     // imm[12|10:5|4:1|11] = inst[31|30:25|11:8|7]
-    return ((instr & 0x80000000) >> 19) as u64
-        | ((instr & 0x80) << 4) as u64 // imm[11]
-        | ((instr >> 20) & 0x7e0) as u64 // imm[10:5]
-        | ((instr >> 7) & 0x1e) as u64; // imm[4:1]
+    return ((instr & 0x80000000) >> 19) as u32
+        | ((instr & 0x80) << 4) as u32 // imm[11]
+        | ((instr >> 20) & 0x7e0) as u32 // imm[10:5]
+        | ((instr >> 7) & 0x1e) as u32; // imm[4:1]
 }
 
-pub fn imm_S(instr: u32) -> u64 {
+pub fn imm_S(instr: u32) -> u32 {
     // imm[11:5] = inst[31:25], imm[4:0] = inst[11:7]
-    return ((instr & 0xfe000000) >> 20) as u64 | ((instr >> 7) & 0x1f) as u64;
+    return ((instr & 0xfe000000) >> 20) as u32 | ((instr >> 7) & 0x1f) as u32;
 }
 
 pub fn imm_I(instr: u32) -> i32 {
@@ -126,15 +126,15 @@ pub fn imm_I(instr: u32) -> i32 {
     return (instr & 0xfff00000) as i32 >> 20;
 }
 
-pub fn imm_U(instr: u32) -> u64 {
+pub fn imm_U(instr: u32) -> u32 {
     // imm[31:12] = inst[31:12]
-    return (instr & 0xfffff000) as u64;
+    return (instr & 0xfffff000) as u32;
 }
 
-pub fn imm_J(instr: u32) -> u64 {
+pub fn imm_J(instr: u32) -> u32 {
     // imm[20|10:1|11|19:12] = inst[31|30:21|20|19:12]
-    return (((instr & 0x80000000) as i32 as i64 >> 11) as u64)// imm[20]
-    | ((instr & 0x3ff00000) >> 20) as u64 // imm[10:1]
-    | ((instr & 0x80000) >> 9) as u64 // imm[11]
-    | (instr & 0xff000) as u64; // imm[19:12]
+    return (((instr & 0x80000000) as i32 >> 11) as u32)// imm[20]
+    | ((instr & 0x3ff00000) >> 20) as u32 // imm[10:1]
+    | ((instr & 0x80000) >> 9) as u32 // imm[11]
+    | (instr & 0xff000) as u32; // imm[19:12]
 }
